@@ -12,47 +12,68 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.12",
-    "@angular/common": "8.2.12",
-    "@angular/compiler": "8.2.12",
-    "@angular/core": "8.2.12",
-    "@angular/forms": "8.2.12",
-    "@angular/platform-browser": "8.2.12",
-    "@angular/platform-browser-dynamic": "8.2.12",
-    "@angular/platform-server": "8.2.12",
-    "@angular/router": "8.2.12",
-    "@nguniversal/module-map-ngfactory-loader": "8.1.1",
+    "@angular/animations": "^12.2.3",
+    "@angular/cdk": "^12.2.3",
+    "@angular/common": "12.2.3",
+    "@angular/compiler": "12.2.3",
+    "@angular/core": "12.2.3",
+    "@angular/forms": "12.2.3",
+    "@angular/material": "^12.2.3",
+    "@angular/material-moment-adapter": "^12.2.3",
+    "@angular/platform-browser": "12.2.3",
+    "@angular/platform-browser-dynamic": "12.2.3",
+    "@angular/platform-server": "12.2.3",
+    "@angular/router": "12.2.3",
+    "@auth0/angular-jwt": "^5.0.2",
+    "@nebular/eva-icons": "^7.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^9.1.2",
+    "@ngtools/webpack": "^1.2.4",
+    "@nguniversal/module-map-ngfactory-loader": "^8.2.6",
+    "@types/googlemaps": "^3.43.3",
+    "angular-calendar": "^0.28.28",
     "aspnet-prerendering": "^3.0.1",
-    "bootstrap": "^4.3.1",
-    "core-js": "^3.3.3",
-    "jquery": "3.4.1",
-    "oidc-client": "^1.9.1",
+    "bootstrap": "^5.1.0",
+    "bootstrap-icons": "^1.5.0",
+    "bootstrap-select": "^1.13.18",
+    "core-js": "^3.16.3",
+    "date-fns": "^2.25.0",
+    "eva-icons": "^1.1.3",
+    "git-commit-msg-linter": "^3.2.8",
+    "git-conventional-commits": "^1.1.0",
+    "hammerjs": "^2.0.8",
+    "jquery": "3.6.0",
+    "material-icons": "^1.2.1",
+    "moment": "^2.29.1",
+    "ngx-auto-unsubscribe": "^3.0.1",
+    "ngx-mask": "^12.0.0",
+    "oidc-client": "^1.11.5",
     "popper.js": "^1.16.0",
-    "rxjs": "^6.5.3",
-    "zone.js": "0.9.1"
+    "rxjs": "^7.3.0",
+    "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.26",
-    "@angular/cli": "^8.3.26",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.12",
-    "@types/jasmine": "~3.4.4",
-    "@types/jasminewd2": "~2.0.8",
-    "@types/node": "~12.11.6",
-    "codelyzer": "^5.2.0",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^5.0.2",
+    "@angular-devkit/build-angular": "^0.1102.8",
+    "@angular/cli": "^12.2.3",
+    "@angular/compiler-cli": "^12.2.3",
+    "@angular/language-service": "^12.2.3",
+    "@schematics/angular": "^12.2.3",
+    "@types/hammerjs": "^2.0.40",
+    "@types/jasmine": "^3.7.8",
+    "@types/jasminewd2": "^2.0.10",
+    "@types/node": "^15.12.5",
+    "codelyzer": "^6.0.2",
+    "jasmine-core": "~3.7.1",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "^6.3.3",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "typescript": "3.5.3"
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "^1.7.0",
+    "typescript": "^4.2.4"
   },
   "optionalDependencies": {
-    "node-sass": "^4.12.0",
-    "protractor": "~5.4.2",
-    "ts-node": "~8.4.1",
-    "tslint": "~5.20.0"
+    "protractor": "~7.0.0",
+    "ts-node": "~10.0.0",
+    "tslint": "~5.20.1"
   }
 }
